@@ -1,64 +1,14 @@
-<template>
-  <div class="header">
-    <div class="nav-bar">
-      <router-link to="/">About me</router-link>
-      <router-link to="/game-projects">games</router-link>
-      <router-link to="/other-projects">Other projects</router-link>
-      <router-link to="/resume">Resume</router-link>
-      <router-link to="/contact">Contact</router-link>
-    </div>
-  </div>
-</template>
-
-<script lang="ts">
-import Vue from "vue";
-
-export default Vue.extend({
-  name: "Header",
-  //   props: {
-  //     msg: String,
-  //   },
-});
+<script setup lang="ts">
+// Navigation component - no logic needed
 </script>
 
-<style scoped lang="less">
-
-@import '../css/variables.less';
-
-.header {
-  width: 100%;
-}
-
-.nav-bar {
-  text-align: right;
-  padding: 20px;
-  line-height: 3em;
-}
-
-a {
-  text-transform: uppercase;
-  margin-left: 15px;
-  margin-right: 15px;
-  padding-bottom: 8px;
-  white-space: nowrap;
-  display: inline-block;
-}
-
-.router-link-exact-active {
-  border: 0px solid @textColor;
-  border-bottom-width: 2px;
-}
-
-@media only screen and (max-width: 620px){
-  .nav-bar {
-    line-height: 2em;
-  }
-
-  a {
-    margin-left: 9px;
-    margin-right: 9px;
-    padding-bottom: 0px;
-  }
-}
-
-</style>
+<template>
+  <header class="header">
+    <nav class="nav-bar">
+      <RouterLink to="/" class="nav-link">About</RouterLink>
+      <RouterLink to="/game-projects" class="nav-link">Games</RouterLink>
+      <RouterLink to="/resume" class="nav-link">Resume</RouterLink>
+      <RouterLink to="/contact" class="nav-link">Contact</RouterLink>
+    </nav>
+  </header>
+</template>
