@@ -1,131 +1,75 @@
-<template>
-    <div>
-      <div class="projects-list">
-        <template v-for="project in projects">
-          <div
-            :key="project.id"
-              @click="showDetails(project)"
-              class="project-item"
-              :class="{ 'wide': project.isWide, 'high': project.isHigh }">
-            <div class="project-item-image" :style="{ 'background-image': 'url(' + project.iconUrl + ')' }">
-            </div>
-            <div class="title-bar" :style="{ 'background-color': project.accentColor + 'DD' }">
-                <div class="title-text">
-                  {{ project.name }}
-                </div>
-              </div>
-          </div>
-        </template>
-      </div>
+<script setup lang="ts">
+import { ref } from 'vue'
+import ProjectDetailsOverlay from '@/components/ProjectDetailsOverlay.vue'
+import type { ProjectData } from '@/data/ProjectData'
 
-      <ProjectDetailsOverlay
-        v-on:close="showPopup = false"
-        :visible="showPopup"
-        :title="popupTitle"
-        :htmlContent="popupContent"
-        :color="popupColor"
-      />
-    </div>
-</template>
+defineProps<{
+  projects: ProjectData[]
+}>()
 
-<script lang="ts">
-import Vue from "vue";
-import ProjectDetailsOverlay from "@/components/ProjectDetailsOverlay.vue";
-import ProjectData from "@/data/ProjectData.ts";
+const showPopup = ref(false)
+const popupTitle = ref('')
+const popupColor = ref('')
+const popupContent = ref('')
 
-export default Vue.extend({
-  name: "ProjectsList",
-  components: {
-    ProjectDetailsOverlay,
-  },
-  props: {
-    projects: Array
-  },
-  data: function () {
-    return {
-      showPopup: false,
-      popupTitle: "",
-      popupColor: "",
-      popupContent: ""
-    };
-  },
-  methods: {
-    showDetails: function (item: ProjectData) {
-      // if (event) {
-      //   alert(event.target);
-      // }
-      this.popupTitle = item.name;
-      this.popupColor = item.accentColor;
-      this.popupContent = item.htmlDescription;
-      this.showPopup = true;
-      window.scrollTo(0,0);
-    },
-  },
-});
+const baseUrl = import.meta.env.BASE_URL
+
+// Resolve image URL with base path for local assets
+function resolveImageUrl(url: string): string {
+  // External URLs (http/https) stay as-is
+  if (url.startsWith('http://') || url.startsWith('https://')) {
+    return url
+  }
+  // Local paths get base URL prepended
+  const cleanUrl = url.startsWith('/') ? url.slice(1) : url
+  return `${baseUrl}${cleanUrl}`
+}
+
+// Fix all local image/video paths in HTML content
+function fixHtmlPaths(html: string): string {
+  // Fix src="/img/..." and src="/d/..." paths
+  return html.replace(/src="\/(?!\/|http)/g, `src="${baseUrl}`)
+}
+
+function showDetails(project: ProjectData) {
+  popupTitle.value = project.name
+  popupColor.value = project.accentColor
+  popupContent.value = fixHtmlPaths(project.htmlDescription)
+  showPopup.value = true
+  window.scrollTo(0, 0)
+}
+
+function closePopup() {
+  showPopup.value = false
+}
 </script>
 
-<style scoped>
+<template>
+  <div>
+    <div class="projects-list">
+      <div
+        v-for="project in projects"
+        :key="project.id"
+        @click="showDetails(project)"
+        class="project-item"
+        :class="{ wide: project.isWide, high: project.isHigh }"
+      >
+        <div
+          class="project-item-image"
+          :style="{ backgroundImage: `url(${resolveImageUrl(project.iconUrl)})` }"
+        />
+        <div class="title-bar">
+          <div class="title-text">{{ project.name }}</div>
+        </div>
+      </div>
+    </div>
 
-.project-item {
-  height: 300px;
-  margin-bottom: 20px;
-  width: 100%;
-  cursor: pointer;
-  position: relative;
-  overflow: hidden;
-}
-
-.project-item-image {
-  background-size: cover;
-  background-position: center;
-  height: 100%;
-  width: 100%;
-  transition: all 0.2s;
-}
-.project-item-image:hover {
-  -webkit-transform: scale(1.1);
-  -ms-transform: scale(1.1);
-  transform: scale(1.1);
-}
-
-.project-item:hover {
-filter: brightness(120%);
-}
-
-.title-bar {
-  position: absolute;
-  bottom: 0px;
-  width: 100%;
-  background-color: #222222;
-}
-
-.title-text {
-  padding: 10px;
-}
-
-@media only screen and (min-width: 620px){
-  .projects-list {
-    max-width: 900px;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-gap: 20px;
-    grid-auto-rows: minmax(250px, auto);
-  }
-
-  .project-item {
-    margin: 0px;
-    height: 100%;
-    width: 100%;
-  }
-
-  .wide {
-    grid-column-end: span 2;
-  }
-  .high {
-    grid-row-end: span 2;
-  }
-}
-
-
-
-</style>
+    <ProjectDetailsOverlay
+      :visible="showPopup"
+      :title="popupTitle"
+      :html-content="popupContent"
+      :color="popupColor"
+      @close="closePopup"
+    />
+  </div>
+</template>
