@@ -1,19 +1,29 @@
-export default class ProjectData {
-    id: string;
-    name: string;
-    htmlDescription: string;
-    iconUrl: string; // used as thumnail
-    isWide: boolean; // thumbnail will take 2 cols in the grid view
-    isHigh: boolean; // thumbnail will take 2 rows in the grid view
-    accentColor: string; // color of title bar 
+export interface ProjectData {
+  id: string
+  name: string
+  htmlDescription: string
+  iconUrl: string
+  isWide: boolean
+  isHigh: boolean
+  accentColor: string
+}
 
-    constructor(id: string, name: string, iconUrl: string, html: string, accentColor = "#000000", isHigh = false, isWide = false){
-        this.id = id;
-        this.name = name;
-        this.htmlDescription = html;
-        this.iconUrl = iconUrl;
-        this.isHigh = isHigh;
-        this.isWide = isWide;
-        this.accentColor = accentColor;
-    }
+export function createProject(
+  id: string,
+  name: string,
+  iconUrl: string,
+  html: string,
+  accentColor = '#8b5cf6',
+  isHigh = false,
+  isWide = false
+): ProjectData {
+  return {
+    id,
+    name,
+    htmlDescription: html,
+    iconUrl,
+    isHigh,
+    isWide,
+    accentColor
+  }
 }
