@@ -1,59 +1,47 @@
+<script setup lang="ts">
+// Contact page
+</script>
+
 <template>
-  <div>
-    
+  <div class="page contact">
     <h1>Let's get in touch</h1>
 
-    <div style="margin-bottom:40px;">I'm looking for a job !<br/>Feel free to contact me about any opportunity, or just to chat about the things I made.</div>
+    <p class="mb-xl">
+      I'm looking for a job! Feel free to contact me about any opportunity, or
+      just to chat about the things I made.
+    </p>
 
-    <ul>
-      <li>
-        <i class="fa fa-user-circle-o fa-lg fa-fw"></i>
-        <a href="http://vishwah.com" target="_blank">https://vishwah.com</a>
+    <ul class="contact-list">
+      <li class="contact-item">
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+          <path d="M4 4h16c1.1 0 2 .9 2 2v12c0 1.1-.9 2-2 2H4c-1.1 0-2-.9-2-2V6c0-1.1.9-2 2-2z"/>
+          <polyline points="22,6 12,13 2,6"/>
+        </svg>
+        <a href="mailto:vishwahsivagurunathan@gmail.com">vishwahsivagurunathan@gmail.com</a>
       </li>
-      <li>
-        <i class="fa fa-envelope-o fa-lg fa-fw"></i>
-        <a href="mailto:vishwahsivagurunathan@gmail.com" target="_blank">Email</a>
+      <li class="contact-item">
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+          <path d="M16 8a6 6 0 0 1 6 6v7h-4v-7a2 2 0 0 0-2-2 2 2 0 0 0-2 2v7h-4v-7a6 6 0 0 1 6-6z"/>
+          <rect x="2" y="9" width="4" height="12"/>
+          <circle cx="4" cy="4" r="2"/>
+        </svg>
+        <a href="https://www.linkedin.com/in/vishwah-sivagurunathan-88940414b/" target="_blank">LinkedIn</a>
       </li>
-      <li>
-        <i class="fa fa-linkedin fa-lg fa-fw" style="vertical-align: -5%;"></i>
-        <a href="https://www.linkedin.com/in/vishwah-sivagurunathan-88940414b/" target="_blank">linkedin</a>
+      <li class="contact-item">
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+          <path d="M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 0 0-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0 0 20 4.77 5.07 5.07 0 0 0 19.91 1S18.73.65 16 2.48a13.38 13.38 0 0 0-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 0 0 5 4.77a5.44 5.44 0 0 0-1.5 3.78c0 5.42 3.3 6.61 6.44 7A3.37 3.37 0 0 0 9 18.13V22"/>
+        </svg>
+        <a href="https://github.com/vishwah13" target="_blank">github.com/vishwah13</a>
       </li>
-      <li>
-        <i class="fa fa-github fa-lg fa-fw"></i>
-        <a href="https://github.com/vishwah13" target="_blank">github.com</a>
-      </li>
-      <li>
-        <i class="fa fa-gamepad fa-lg fa-fw"></i>
-        <a href="https://vishwah13.itch.io/" target="_blank">itch.io</a>
+      <li class="contact-item">
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+          <line x1="6" y1="3" x2="6" y2="15"/>
+          <circle cx="18" cy="6" r="3"/>
+          <circle cx="6" cy="18" r="3"/>
+          <path d="M18 9a9 9 0 0 1-9 9"/>
+        </svg>
+        <a href="https://vishwah13.itch.io/" target="_blank">vishwah13.itch.io</a>
       </li>
     </ul>
-
   </div>
 </template>
-
-<style scoped>
-ul {
-  list-style: none;
-  padding-inline-start: 0px;
-}
-
-li {
-    line-height: 1.8em;
-    font-size: 1.2em;
-  }
-
-@media only screen and (min-width: 620px){
-  ul {
-    padding-inline-start: 40px;
-  }
-
-  li {
-    line-height: 2em;
-    font-size: 1.4em;
-  }
-}
-
-i {
-  margin-right: 20px;
-}
-</style>
